requirements.txt
python-dotenv==1.0.1
requests==2.31.0
EyeSim-Python==1.2.4
numpy
pandas
openai
transformers
huggingface_hub
torch==2.2.0
accelerate
httpx
tqdm